Match the default module in hierarchical names
When a dependency is specified in the PackageManifest and the DependencyManifest, the module details are not added to the BlendedManifest as we pick the details from the PackageManifest. As a result, when the module is of hierarchical nature, the resolver fails to pick the dependency from the blended manifest as it tries to check if the module is one of the modules of the dependency recorded.